+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Decreased message spacing for messages sent in quick succession.
 - Improved message loading spinner.
+- Ported to the Adwaita shortcuts dialog, instead of the deprecated GTK one.
 
 ## [0.18.8] - 2026-03-03